Cleaned Up Code
The variables did not have a descriptive and easy to understand naming convention. Many of the variable names were updated

The code was not adaptable / dynamic to different type of boards (eg, boards that may have a different pin structure, or boards that have 1 layer or 5 layers in the future. I have made initial updates to fix this.

The water level sensor must use an internal pull-up resistor. I have modified the code for this. Further information can be found by researching internal pullup resistors esp32

The sensors will not read accurate values if the atomizers are on. The read water level function was cleaned up and is only called when the atomizers are off

The code is not properly commented. I have begun fixing this.

There should be separation between board configuration, location configuration, etc. so I moved serial number to credentials (location configuration)

Fixed naming conventions